chore: address more review comments
