+++ .gitignore
@@ -64,4 +64,4 @@ override.tf.json
 .terraformrc
 terraform.rc
 .env
-internal_users.yml
+/tests/unit/resources/config/opensearch-security/internal_users.yml
